+++ lib/db_mod/statements/configuration/single.rb
@@ -41,7 +41,7 @@ module DbMod
         }
 
         # Extend the given method definition with additional
-        # result coercion, if specified.
+        # result coercion, if specified using {MethodConfiguration#single.
         #
         # @param definition [Proc] base method definition
         # @param config [MethodConfiguration] method configuration